requirements.txt
anyio==2.2.0
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
black==20.8b1
bleach==3.3.0
boto3==1.17.57
botocore==1.20.57
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
clarifai==2.6.2
click==7.1.2
colorama==0.4.4
configparser==3.8.1
confuse==1.4.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
descartes==1.1.0
EasyProcess==0.3
emoji==0.6.0
entrypoints==0.3
fastavro==0.22.13
feast==0.10.2
future==0.15.0
google-api-core==1.26.3
google-auth==1.29.0
google-cloud-bigquery==2.13.1
google-cloud-bigquery-storage==2.4.0
google-cloud-core==1.4.4
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.9.2
jupyter-server==1.6.4
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
kaggle==1.5.12
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
libcst==0.3.18
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
MeaningCloud-python==2.0.0
metaflow==2.2.10
missingno==0.4.2
mistletoe==0.7.2
mistune==0.8.4
mizani==0.7.3
mmh3==3.0.0
mypy-extensions==0.4.3
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numpy==1.19.5
packaging==20.9
palettable==3.3.0
pandas==1.2.4
pandas-profiling==2.11.0
pandavro==1.5.2
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
patsy==0.5.1
pexpect==4.8.0
phik==0.11.2
pickleshare==0.7.5
Pillow==8.2.0
plotnine==0.8.0
plyer==2.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
Pygments==2.8.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-slugify==4.0.1
python-telegram-bot==13.1
python3-xlib==0.15
pytz==2020.5
PyVirtualDisplay==2.0
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
statsmodels==0.12.2
tabulate==0.8.9
tangled-up-in-unicode==0.0.7
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.1
tomlkit==0.7.0
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
urllib3==1.26.4
visions==0.6.0
wcwidth==0.2.5
webdriverdownloader==1.1.0.3
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
